@@ -10,26 +10,24 @@
   "bugs": "https://github.com/eight04/textarea-plus/issues",
   "license": "MIT",
   "author": "eight04 <eight04@gmail.com>",
-  "main": "textarea-plus.user.js",
   "repository": "eight04/textarea-plus",
   "scripts": {
-    "test": "eslint **/*.js && mocha",
+    "test": "eslint --ext .mjs . && mocha",
     "preversion": "npm test",
     "version": "npm run build && git add .",
     "postversion": "git push --follow-tags",
-    "build": "npm run build-lib && npm run build-userscript",
-    "build-lib": "rollup -c",
-    "build-userscript": "inlinejs src/textarea-plus.user.js -o dist/textarea-plus.user.js"
+    "build": "rollup -c"
   },
-  "private": true,
+  "eslintIgnore": [
+    "dist"
+  ],
   "devDependencies": {
-    "chai": "^4.1.2",
-    "eslint": "^5.16.0",
-    "inline-js": "^0.6.0",
-    "mocha": "^5.0.4",
-    "rollup": "^0.57.0",
-    "rollup-plugin-commonjs": "^9.1.0",
-    "userscript-meta-cli": "^0.3.0"
+    "chai": "^4.3.4",
+    "eslint": "^8.2.0",
+    "mocha": "^9.1.3",
+    "rollup": "^2.60.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "userscript-meta-cli": "^0.4.2"
   },
   "userscript": {
     "name": "Textarea Plus",
